requirements.txt
dbt-core
pandas
numpy
constants
apache-airflow
apache-airflow-providers-mongo